Chore

+++ issuetype.go
@@ -1,7 +1,7 @@
 package backlog
 
 type IssueType struct {
-	Id           int    `json:"id"`
+	Id           uint64 `json:"id"`
 	ProjectId    int    `json:"projectId"`
 	Name         string `json:"name"`
 	Color        string `json:"color"`
